pyproject.toml
[project]
name = "showmethemoney"
version = "0.1.0"
description = "ETF 및 주식 종목 데이터를 수집, 저장, 시각화하는 대시보드 애플리케이션"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic-settings",
    "python-dotenv",
    "httpx",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "azure-core-tracing-opentelemetry",
    "azure-monitor-opentelemetry",
    "azure-monitor-opentelemetry-exporter",
    "opentelemetry-instrumentation-httpx",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-instrumentation-openai",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "black",
    "ruff",
    "beautifulsoup4>=4.12.3",
    "msgpack>=1.1.0",
    "apscheduler>=3.10.4",
]

[tool.uv]
managed = true
dev-dependencies = []

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]

[tool.hatch.build.targets.wheel]
packages = ["backend/app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
